test: Sections 테스트 추가
